Added ability to tweet Wikipedia links; addressed name issues
The bot will now check to see if the pulsar has a Wikipedia page; if so, it appends a link to the tweet's body. I've also fixed an issue regarding names. In entries with a PSRB field (epoch 1950), it's listed before the PSRJ field (epoch 2000), so I should have been explicit about which name is compared to the input name. This also helps if a pulsar has a Wikipedia article under its SPRJ name and for some reason there's no redirect from its PSRB name.