Added flux densities.
I've added in flux densities, with values at 400 MHz and 1400 MHz being prioritized and others used if values at those frequencies are not available. The tweets are also getting rather long, so I've made Wikipedia links optional if that would put it over. Perhaps it would be worth adding a second tweet in that scenario, because more changes will make it more likely that the output will be pushed over the 280-character threshold.